apps/trucking-agent/pyproject.toml
[project]
name = "data-agent-spike"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework-core>=1.0.1",
    "agent-framework-foundry>=1.0.1",
    "agent-framework-openai>=1.0.1",
    "aiohttp>=3.13.5",
    "azure-identity>=1.25.3",
    "fastapi>=0.136.0",
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn>=0.44.0",
]
